scripts/python/requirements.txt
python-dotenv
psycopg2